+++ spring-integration-cloudevents/src/main/java/org/springframework/integration/cloudevents/transformer/strategies/package-info.java
@@ -1,0 +1,3 @@
+
+@org.jspecify.annotations.NullMarked
+package org.springframework.integration.cloudevents.transformer.strategies;
