--- spring-integration-cloudevents/src/main/java/org/springframework/integration/cloudevents/v1/transformer/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Base package for CloudEvents transformer support.
- */
-
-@org.jspecify.annotations.NullMarked
-package org.springframework.integration.cloudevents.v1.transformer;
