--- spring-integration-cloudevents/src/main/java/org/springframework/integration/cloudevents/v1/transformer/utils/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Base package for CloudEvents  transformer util support.
- */
-
-@org.jspecify.annotations.NullMarked
-package org.springframework.integration.cloudevents.v1.transformer.utils;
