+++ src/reference/antora/modules/ROOT/nav.adoc
@@ -62,7 +62,6 @@
 ** xref:logging-adapter.adoc[]
 ** xref:functions-support.adoc[]
 ** xref:kotlin-functions.adoc[]
-* xref:cloudevents/cloudevents-transform.adoc[]
 * xref:dsl.adoc[]
 ** xref:dsl/java-basics.adoc[]
 ** xref:dsl/java-channels.adoc[]
@@ -125,6 +124,7 @@
 ** xref:amqp/amqp-1.0.adoc[]
 * xref:camel.adoc[]
 * xref:cassandra.adoc[]
+* xref:cloudevents-transform.adoc[]
 * xref:debezium.adoc[]
 * xref:event.adoc[]
 * xref:feed.adoc[]
